♻️ refactor(tools): route node mutations through tool service
